+++ kmp-maps-compose/src/webMain/kotlin/eu/buney/maps/GroundOverlay.web.kt
@@ -43,6 +43,7 @@ actual fun GroundOverlay(
                         clickable = clickable,
                     )
                 ),
+                image = image,
                 bounds = bounds,
                 bearing = bearing,
                 transparency = transparency,